requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0,<3.0.0  # Pin to 2.x to support dataset scripts
evaluate>=0.4.0
sentence-transformers>=2.2.0
sacrebleu>=2.3.0
unbabel-comet>=2.2.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
accelerate>=0.20.0
